Added stdbool.h, fixed uint => unsigned int, size_t => int

+++ uppgifter/fas1/sprint2/lager/common.h
@@ -1,6 +1,8 @@
 #ifndef __common_h__
 #define __common_h__
 
+#include <stdbool.h>
+
 typedef union element elem_t;
 
 /// Element wrapper
@@ -10,16 +12,16 @@ typedef union element elem_t;
 ///
 union element
 {
-  int   i; 
-  uint  u;
-  void *p; 
-  float f;
+  int           i; 
+  unsigned int  u;
+  void         *p; 
+  float         f;
 };
 
 /// Functions for handling genericity
 typedef elem_t(*element_copy_fun)(elem_t);
 typedef void(*element_free_fun)(elem_t);
-typedef size_t(*element_comp_fun)(elem_t, elem_t);
+typedef int(*element_comp_fun)(elem_t, elem_t);
 typedef bool(*elem_apply_fun)(elem_t elem, void *data);
 
 #endif
